@@ -13,26 +13,30 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~8.1.2",
-    "@angular/compiler": "~8.1.2",
-    "@angular/core": "~8.1.2",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "~8.2.14",
+    "@angular/core": "^8.2.14",
     "@angular/fire": "^5.2.1",
-    "@angular/forms": "~8.1.2",
-    "@angular/platform-browser": "~8.1.2",
-    "@angular/platform-browser-dynamic": "~8.1.2",
-    "@angular/router": "~8.1.2",
+    "@angular/forms": "~8.2.14",
+    "@angular/platform-browser": "~8.2.14",
+    "@angular/platform-browser-dynamic": "~8.2.14",
+    "@angular/router": "~8.2.14",
+    "@angular/service-worker": "^8.2.14",
     "@ionic-native/camera": "^5.15.1",
     "@ionic-native/core": "^5.0.0",
     "@ionic-native/device": "^5.15.0",
-    "@ionic-native/fcm": "^5.15.0",
+    "@ionic-native/fcm": "^5.16.0",
     "@ionic-native/file": "^5.15.1",
     "@ionic-native/file-path": "^5.15.1",
-    "@ionic-native/firebase": "^5.15.0",
-    "@ionic-native/geolocation": "^5.15.0",
+    "@ionic-native/firebase": "^4.20.0",
+    "@ionic-native/geolocation": "^5.16.0",
     "@ionic-native/ionic-webview": "^5.15.1",
-    "@ionic-native/native-geocoder": "^5.15.0",
-    "@ionic-native/splash-screen": "^5.0.0",
+    "@ionic-native/native-geocoder": "^5.16.0",
+    "@ionic-native/social-sharing": "^5.17.1",
+    "@ionic-native/splash-screen": "^5.17.1",
     "@ionic-native/status-bar": "^5.0.0",
+    "@ionic-native/toast": "^5.16.0",
+    "@ionic-native/unique-device-id": "^5.16.0",
     "@ionic/angular": "^4.7.1",
     "@ionic/storage": "^2.2.0",
     "angular-ng-autocomplete": "^1.1.14",
