@@ -1,4 +1,4 @@
-//Angular
+// Angular
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { RouteReuseStrategy } from '@angular/router';
@@ -9,17 +9,18 @@ import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
 import { FormsModule } from '@angular/forms';
 import { ReactiveFormsModule } from '@angular/forms';
-import { HttpClient, HttpHeaders } from '@angular/common/http';
+// Environments
 import { environment } from './environments/environment';
 import { HttpClientModule } from '@angular/common/http';
+// Imported Plugin
 import { Device } from '@ionic-native/device/ngx';
 import { Geolocation } from '@ionic-native/geolocation/ngx';
-
+import { UniqueDeviceID } from '@ionic-native/unique-device-id/ngx';
+// Services
 import { AlldciService } from './services/alldci.service';
 import { DataService } from './services/item.service';
-import { NativeGeocoder } from '@ionic-native/native-geocoder/ngx';
-
 import { NotificationsService } from './tab1/notification.service';
+import { ResultatVerificationService, User } from './services/verifcode.service';
 import { AuthService } from './services/AuthService';
 import { TextSearchService } from './services/textsearch.service';
 import { ResultatInteractionService } from './services/resultatInteraction.service';
@@ -28,23 +29,27 @@ import { VilleService } from './services/city.service';
 import { GaleniqueService } from './services/galenique.service';
 import { AnatomiqueService } from './services/anatomique.service';
 import { QryCustomerService } from './services/customers.service';
+import { QryQuotationsService } from './services/quotation.services';
+import { QryStatNotifs } from './tab1/statnotif.service';
+import { ManageUsers } from './services/manageUsers.service';
 
+// Captcha
+
+import { QrySignalementService } from './services/signalement.service';
 import { InteractionSearch } from './services/interaction.service';
 import { InteractionComponent } from './tab3/interaction/interaction.component';
 import { AuthenticateService } from './services/authentication.service';
 import { SchoolsService } from './services/school.service';
-// tslint:disable-next-line: import-spacing
 import { QryInteractionService } from './services/Qryinteraction.service';
-
+import { DentalService } from './dentiste/dental.service';
 // Firebase modules
 import { AngularFireModule } from '@angular/fire';
-import * as firebase from 'firebase';
-import { Firebase } from '@ionic-native/firebase/ngx';
 import { AngularFirestoreModule } from '@angular/fire/firestore';
 import { AngularFireDatabase, AngularFireList } from '@angular/fire/database';
-
+import { FCM } from '@ionic-native/fcm/ngx';
 import { Camera, CameraOptions } from '@ionic-native/camera/ngx';
 import { File } from '@ionic-native/file/ngx';
+import { IonicStorageModule } from '@ionic/storage';
 
 @NgModule({
   declarations: [AppComponent],
@@ -57,8 +62,7 @@ import { File } from '@ionic-native/file/ngx';
     ReactiveFormsModule,
     AngularFirestoreModule, // Only required for database features
     AngularFireModule.initializeApp(environment.firebase),
-
-    IonicModule.forRoot(), AppRoutingModule],
+    IonicModule.forRoot(), IonicStorageModule.forRoot(), AppRoutingModule],
   providers: [
     StatusBar,
     AuthService,
@@ -67,20 +71,27 @@ import { File } from '@ionic-native/file/ngx';
     AnatomiqueService,
     TextSearchService,
     NotificationsService,
+    DentalService,
     QryCustomerService,
+    QryStatNotifs,
+    QryQuotationsService,
     Device,
     Geolocation,
-    NativeGeocoder,
     AuthenticateService,
+    ManageUsers,
+    ResultatVerificationService,
     SchoolsService,
     InteractionComponent,
     InteractionSearch,
     QryInteractionService,
     ResultatInteractionService,
+    QrySignalementService,
     DciSearch,
     VilleService,
     AlldciService,
+    UniqueDeviceID,
     AngularFireDatabase,
+    FCM,
     Camera,
     File,
     SplashScreen,
