@@ -3,7 +3,7 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  production: false,
+  production: true,
   firebase: {
     apiKey: 'AIzaSyAhdaQTRQLYIHfGC7gQpxufsWVVh93bQjI',
     authDomain: 'woven-art-587.firebaseapp.com',
