@@ -7,13 +7,34 @@ import { SchoolsService } from '../services/school.service';
 import { Geolocation } from '@ionic-native/geolocation/ngx';
 import { NativeGeocoder, NativeGeocoderOptions } from '@ionic-native/native-geocoder/ngx';
 import { map } from 'rxjs/operators';
+import { Storage } from '@ionic/storage';
+import { ToastController } from '@ionic/angular';
 @Component({
   selector: 'app-login',
   templateUrl: './login.component.html',
   styleUrls: ['./login.component.scss'],
 })
 export class LoginComponent implements OnInit {
+  constructor(
+
+    private navCtrl: NavController,
+    public authService: AuthenticateService,
+    private formBuilder: FormBuilder,
+    private geolocation: Geolocation,
+    public storage: Storage,
+    public toastController: ToastController,
+    private schoolsService: SchoolsService
+
+  ) {  // set a key/value
+    storage.set('name', 'Sexy');
+
+    // Or to get a key/value pair
+    storage.get('age').then((val) => {
+      console.log('Your age is', val);
+    });
+  }
   schools: any;
+  userState: any;
   geoLatitude: number;
   geoLongitude: number;
   geoAccuracy: number;
@@ -22,28 +43,13 @@ export class LoginComponent implements OnInit {
   watchLocationUpdates: any;
   loading: any;
   isWatching: boolean;
-
+  userEmail: string;
   // Geocoder configuration
-  geoencoderOptions: NativeGeocoderOptions = {
-    useLocale: true,
-    maxResults: 5
-  };
   device: any;
   logged: boolean;
-  constructor(
-
-    private navCtrl: NavController,
-    private authService: AuthenticateService,
-    private formBuilder: FormBuilder,   
-    private geolocation: Geolocation,
-    private schoolsService: SchoolsService,
-    private nativeGeocoder: NativeGeocoder
-
-  ) { }
-
-
   // tslint:disable-next-line: variable-name
   validations_form: FormGroup;
+  // tslint:disable-next-line: member-ordering
   errorMessage = '';
 
   // tslint:disable-next-line: variable-name
@@ -57,7 +63,13 @@ export class LoginComponent implements OnInit {
       { type: 'minlength', message: 'Votre mot de passe doit contenir au moins 5 caractères.' }
     ]
   };
-
+  async logoutToast() {
+    const toast = await this.toastController.create({
+      message: 'Vous êtes deconnecté.',
+      duration: 2000
+    });
+    toast.present();
+  }
   ngOnInit() {
     this.schoolsService.getSchool().snapshotChanges().pipe(
       map(changes =>
@@ -67,12 +79,10 @@ export class LoginComponent implements OnInit {
       )
     ).subscribe(schools => {
       this.schools = schools;
-     // console.log(schools);
+      // console.log(schools);
     });
-   // console.log('Device UUID is: ' + this.device.uuid);
-
-
-    //Validations patterns
+    // console.log('Device UUID is: ' + this.device.uuid);
+    // Validations patterns
     this.validations_form = this.formBuilder.group({
       email: new FormControl('', Validators.compose([
         Validators.required,
@@ -84,24 +94,35 @@ export class LoginComponent implements OnInit {
       ])),
     });
   }
- // Get current coordinates of device
- getGeolocation() {
-  this.geolocation.getCurrentPosition().then((resp) => {
-    this.geoLatitude = resp.coords.latitude;
-    this.geoLongitude = resp.coords.longitude;
-    this.geoAccuracy = resp.coords.accuracy;
-    console.log(this.geoLatitude);
-   // this.getGeoencoder(this.geoLatitude,this.geoLongitude);
-  // }).catch((error) => {
-    // alert('Error getting location'+ JSON.stringify(error));
-   });
-}
-
+  // Get current coordinates of device
+  getGeolocation() {
+    this.geolocation.getCurrentPosition().then((resp) => {
+      this.geoLatitude = resp.coords.latitude;
+      this.geoLongitude = resp.coords.longitude;
+      this.geoAccuracy = resp.coords.accuracy;
+      console.log(this.geoLatitude);
+      // this.getGeoencoder(this.geoLatitude,this.geoLongitude);
+      // }).catch((error) => {
+      // alert('Error getting location'+ JSON.stringify(error));
+    });
+   
+  }
   loginUser(value) {
     this.authService.loginUser(value)
       .then(res => {
         this.errorMessage = '';
         this.logged = true;
+    
+        this.navCtrl.navigateForward('tabs/tab1');
+      }, err => {
+        this.errorMessage = err.message;
+      });
+  }
+
+  resetPassword(value) {
+    this.authService.resetPassword(value)
+      .then(res => {
+        this.errorMessage = '';
         this.navCtrl.navigateForward('tabs/tab1');
       }, err => {
         this.errorMessage = err.message;
@@ -113,11 +134,11 @@ export class LoginComponent implements OnInit {
       this.navCtrl.navigateForward('tabs/tab1');
     });
   }
-
   goToRegisterPage() {
     this.navCtrl.navigateForward('/account');
   }
   goToRegisterPageStudent() {
     this.navCtrl.navigateForward('/students');
   }
+  
 }
