@@ -1,7 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { ModalController } from '@ionic/angular';
-import { NavController } from '@ionic/angular';
 import { ModalPagePage } from './modal-page/modal-page.page';
+import { AuthService } from './../services/AuthService';
+import { ConnectionService } from 'ng-connection-service';
+
 @Component({
   selector: 'app-tab3',
   templateUrl: './tab3.page.html',
@@ -9,20 +11,42 @@ import { ModalPagePage } from './modal-page/modal-page.page';
 })
 export class Tab3Page implements OnInit {
   dataFromModal;
+  interactions: any;
+  userEmail: string;
+  userName: string;
 
-  constructor(private modalController: ModalController) {}
+  status = '';
+  isConnected = true;
+  navCtrl: any;
+  logged: any;
+  constructor(private modalController: ModalController,
+              public authService: AuthService,
+              public connectionService: ConnectionService) { }
 
-  ngOnInit() { }
+  ngOnInit() {
+    // tslint:disable-next-line: no-unused-expression
+    this.userEmail;
+    // tslint:disable-next-line: no-unused-expression
+    this.logged;
 
-  async settings() {
-    const modal = await this.modalController.create({
-      component: ModalPagePage,
-      componentProps: { website: 'edupala.com' },
-      backdropDismiss: false,
+    this.connectionService.monitor().subscribe(isConnected => {
+      this.isConnected = isConnected;
+      if (this.isConnected) {
+        this.status = 'Connecté à Internet';
+        console.log(this.status);
+      } else {
+        this.status = 'Pas de connection Internet';
+        console.log(this.status);
+      }
     });
-
-    modal.present();
-    this.dataFromModal = await modal.onWillDismiss();
+    if (this.authService.userDetails()) {
+      this.userEmail = this.authService.userDetails().email;
+      console.log(this.userEmail);
+      this.logged = true;
+      this.userName = this.authService.userDetails().displayName;
+      console.log(this.userName);
+    } else {
+      this.logged = false;
+    }
   }
-
 }
